Backscatter transport with cumulative slabs for correct depth data
