+++ HZETRN/tag_utils.py
@@ -1,18 +1,4 @@
-import subprocess
 import numpy as np
-
-try:
-    from scipy.spatial import ConvexHull
-except ImportError:
-    raise ImportError("scipy.spatial.ConvexHull could not be imported.")
-
-try:
-    from itaps import iMesh, iBase
-
-    HAVE_PYTAPS = True
-except ImportError:
-
-    HAVE_PYTAPS = False
 
 try:
    from pyne import material
@@ -20,6 +6,10 @@ try:
 except:
     raise ImportError("The PyNE dependencies could not be imported.")    
 
+try:
+    from itaps import iMesh, iBase
+except ImportError:
+    raise ImportError("Could not import iMesh and/or iBase.")
 
 """
 Produce length-one tuples randomly oriented on a sphere
@@ -41,131 +31,6 @@ def get_rand_dirs(number):
 	    rays.append([x, y, z])
     return np.array(rays)
 
-def get_verts_on_sphere(xyz):
-    """
-    Order the set of points as a triangular mesh 
-    xyz	     nx3 np array with points on sphere
-    return   triangular facets as sets of points
-    """
-    hull = ConvexHull(xyz)
-    indices = hull.simplices
-    return xyz[indices]
-
-def create_meshed_sphere(facets, scale):
-    """
-    facets nx3x3 array of triangulated points on a sphere
-    """
-    msph = iMesh.Mesh()
- 
-    for facet in facets:
-	# ToDo:  is it possible to createVtx with vtxs, i.e.
-	#        an array of facets?
-	verts = msph.createVtx(facet*scale)
-
-	tri, stat = msph.createEntArr(iMesh.Topology.triangle, verts)
-	
-    return msph
-
-def create_tagged_meshed(data, facets, scale):
-    """
-    data   data indexed the same as the facets
-    facets nx3x3 array of triangulated points on a sphere
-    """
-    msph = iMesh.Mesh()
- 
-    for facet in facets:
-	# ToDo:  is it possible to createVtx with vtxs, i.e.
-	#        an array of facets?
-	verts = msph.createVtx(facet*scale)
-
-	tri, stat = msph.createEntArr(iMesh.Topology.triangle, verts)
-	
-    return msph
-
-def mesh_sphere(xyz, scale):
-    """
-    xyz	 set of points on a sphere: nx3x3
-    """
-    facets = get_verts_on_sphere(xyz)
-    return create_meshed_sphere(facets, scale)
-
-
-def mesh_write(filename, numOfPts):
-    xyz = get_rand_dirs(numOfPts)
-    mesh = mesh_sphere(xyz, 1.0)
-    mesh.save(filename)
-    return 
-
-def get_cols_from_file(infile, data_column):
-    "Create a scaled, tagged mesh from data in a file."
-    # extract direction and one data column
-    # us columns rather than fields to count.
-    field_width = 13
-    col_start = str((data_column-1)*field_width)
-    col_end   = str(data_column*field_width)
-    # Columns 1-39 contain the three direction values
-    column_list = "1-39,{}-{}".format(col_start, col_end)
-    print "cut column values", column_list
-    # Call a linux command to get desired columns; capture output
-    args = ["cut", "-c", column_list, infile]
-    p = subprocess.Popen(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    output, err = p.communicate()
-    all_lines = output.splitlines()
-    for line in all_lines:
-        print line
-    # Skip first (header) and last two (statistics) lines
-    lines = all_lines[1:len(all_lines)-2]
-    return lines
-
-"""
-    Read certain columns from the infile.
-    Using direction vectors on each line of the file, and data
-    from dat_column, create a tagged iMesh.Mesh that can be
-    directly writen to a *.vtk file for viewing with VisIt.
-    The direction vectors are expected to be unit length and will
-    be scaled by scale.
-"""
-def tag_mesh(infile, data_column, scale):
-    "Create a scaled, tagged mesh from data in a file."
-
-    lines = get_cols_from_file(infile, data_column)
-    vals = []
-    for line in lines:
-        vals.append(map(float, line.split()))
-    vtxd = np.array(vals)
-    
-    # Separate into vertexes (1st 3) and data
-    vtcs = vtxd[:,:3]
-    data = vtxd[:,3]
-
-    # Triangulate 
-    hull       = ConvexHull(vtcs)
-    indices    = hull.simplices
-    facets     = vtcs[indices]
-    di         = data[indices]
-    num_facets = facets.shape[0]
-
-    # Create the mesh, and a data tag on it
-    msph = iMesh.Mesh()
-    point_data = msph.createTag("point_data", 1, float)
-
-    # Use indexing to get matching data at each vertex
-    for i in range(num_facets):
-        facet        = facets[i]
-	data_at_vtcs = di[i]
-	# Create an entity handle for each point in the facet
-	verts = msph.createVtx(facet*scale)
-	# Tag each entity handle (lhs key) with 
-	# corresponding data value (rhs)
-	point_data[verts[0]] = data_at_vtcs[0]
-	point_data[verts[1]] = data_at_vtcs[1]
-	point_data[verts[2]] = data_at_vtcs[2]
-
-	# Tell the mesh about this triangular facet
-	tri, stat = msph.createEntArr(iMesh.Topology.triangle, verts)
-	
-    return msph
-    
 
 """
 function to transform the tags into strings
@@ -244,6 +109,6 @@ def get_fnames_for_vol(path):
     for vol in vol_mat_dict:
         if 'graveyard' in vol_mat_dict[vol].lower():
 	   vol_fname_dict[vol] = 'graveyard' 
-	   print 'Setting name for vol', vol, 'graveyard'
+	   # print 'Setting name for vol', vol, 'graveyard'
     return vol_fname_dict
